requirements.txt
fastapi
uvicorn[standard]
python-multipart
pillow
pytesseract
opencv-python-headless
pydantic
pydantic-settings
SQLAlchemy>=2.0
pandas
numpy
python-dateutil
rapidfuzz
python-dotenv